+++ build.gradle
@@ -86,6 +86,11 @@ dependencies {
     }
 }
 
+mixin {
+    defaultObfuscationEnv searge
+    add sourceSets.main, 'mixins.toastclient.refmap.json'
+}
+
 processResources {
     // this will ensure that this task is redone when the versions change.
     inputs.property 'version', project.version
