+++ src/main/java/me/remainingtoast/toastclient/gui/click/modules/frame/Frame.java
@@ -1,6 +1,6 @@
-package me.remainingtoast.toastclient.gui.click.panel.modules.frame;
+package me.remainingtoast.toastclient.gui.click.modules.frame;
 
-import me.remainingtoast.toastclient.gui.click.panel.modules.Component;
+import me.remainingtoast.toastclient.gui.click.modules.Component;
 import me.remainingtoast.toastclient.util.RenderUtil;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.ScaledResolution;
