--- src/main/java/me/remainingtoast/toastclient/gui/click/panel/panels/HudEditorPanel.java
@@ -1,11 +1,0 @@
-package me.remainingtoast.toastclient.gui.click.panel.panels;
-
-import me.remainingtoast.toastclient.gui.click.panel.Panel;
-
-public class HudEditorPanel extends Panel {
-
-    public HudEditorPanel() {
-        super("Hud Editor");
-    }
-
-}
