+++ src/main/java/me/remainingtoast/toastclient/gui/click/panels/HudEditorPanel.java
@@ -1,0 +1,11 @@
+package me.remainingtoast.toastclient.gui.click.panels;
+
+import me.remainingtoast.toastclient.gui.click.Panel;
+
+public class HudEditorPanel extends Panel {
+
+    public HudEditorPanel() {
+        super("Hud Editor");
+    }
+
+}
