+++ src/main/java/me/remainingtoast/toastclient/module/modules/gui/Console.java
@@ -1,6 +1,6 @@
 package me.remainingtoast.toastclient.module.modules.gui;
 
-import me.remainingtoast.toastclient.gui.click.panel.modules.frame.ConsoleFrame;
+import me.remainingtoast.toastclient.gui.click.modules.frame.ConsoleFrame;
 import me.remainingtoast.toastclient.module.Module;
 import me.remainingtoast.toastclient.module.ModuleManifest;
